Some nits
